+++ lib/schemacop/field_node.rb
@@ -16,11 +16,33 @@ module Schemacop
     def validate(data, collector)
       unless data.key?(name)
         collector.error "Missing key #{name.inspect}." if @required
-        return
       end
-      collector.path "/#{name}", name do
-        super(data[name], collector)
+
+      collector.path "/#{name}", name, :hash do
+        value, default_applied = apply_default!(data[name], collector)
+
+        unless data.key?(name) || default_applied
+          next
+        end
+
+        super(value, collector)
       end
+    end
+
+    private
+
+    def apply_default!(data, collector)
+      return data, false unless data.nil?
+
+      @types.each do |type|
+        if type.option?(:default)
+          default = type.option(:default)
+          collector.override_value(default)
+          return default, true
+        end
+      end
+
+      return nil, false
     end
   end
 end